Update dependencies in go.mod and go.sum, refactor RISKEN MCP server context handling, and add unit tests for RISKEN client retrieval
